Rename optuna module to hpo
